Organize packages for CreateLao examples
